+++ src/index.js
@@ -1,4 +1,4 @@
-export {render} from "./render.js";
+export {app} from "./app.js";
 export {svg} from "./svg.js";
 export {state} from "./state.js";
 export {constrain} from "./math.js";
