--- src/mark.js
@@ -1,8 +1,0 @@
-export class Mark {
-  constructor(data, options, children) {
-    this._data = data;
-    this._options = options;
-    this._nodes = [];
-    this._children = children;
-  }
-}
