+++ test/apps/index.js
@@ -1,5 +1,5 @@
 import * as cm from "../../src/index.js";
 
 export function circle() {
-  return cm.render({draw: [cm.svg("circle", {cx: 50, cy: 50, r: 50})]}).node();
+  return cm.app({draw: [cm.svg("circle", {cx: 50, cy: 50, r: 50})]}).render();
 }
