+++ test/svg.spec.js
@@ -1,7 +1,7 @@
 import {svg} from "../src/index.js";
 import {test, expect} from "vitest";
 
-test("svg(tag, options) should return a SVG collection with expcted attributes", () => {
+test("svg(tag, options) should return a SVG mark with expcted attributes", () => {
   const s = svg("circle", {cx: 0, cy: 0, r: 10});
   expect(s._tag).toBe("circle");
   expect(s._data).toEqual([0]);
@@ -10,7 +10,7 @@ test("svg(tag, options) should return a SVG collection with expcted attributes",
   expect(s._nodes).toEqual([]);
 });
 
-test("svg(tag, data, options) should return a SVG collection with expcted attributes", () => {
+test("svg(tag, data, options) should return a SVG mark with expcted attributes", () => {
   const s = svg("circle", [1, 2, 3], {cx: 0, cy: 0, r: 10});
   expect(s._tag).toBe("circle");
   expect(s._data).toEqual([1, 2, 3]);
